Handle randomizing matrix question order
