+++ src/components/response/MatrixInput.tsx
@@ -2,13 +2,14 @@ import {
   Box, Flex, Radio, Text, Checkbox,
 } from '@mantine/core';
 import {
-  ChangeEvent, useEffect, useMemo, useState,
+  ChangeEvent, useMemo,
 } from 'react';
 import { MatrixResponse, StringOption } from '../../parser/types';
 import { ReactMarkdownWrapper } from '../ReactMarkdownWrapper';
 import { useStoreDispatch, useStoreActions } from '../../store/store';
 import checkboxClasses from './css/Checkbox.module.css';
 import radioClasses from './css/Radio.module.css';
+import { useStoredAnswer } from '../../store/hooks/useStoredAnswer';
 
 function CheckboxComponent({
   _choices,
@@ -22,7 +23,7 @@ function CheckboxComponent({
   _choices: StringOption[],
   _n: number,
   idx: number,
-  question: StringOption,
+  question: string,
   answer: { value: Record<string, string> },
   onChange: (event: ChangeEvent<HTMLInputElement>, questionKey: string, option: StringOption) => void
   disabled: boolean
@@ -40,8 +41,8 @@ function CheckboxComponent({
         <Checkbox
           disabled={disabled}
           key={`${checkbox.label}-${idx}`}
-          checked={answer.value[question.label].split('|').includes(checkbox.value)}
-          onChange={(event) => onChange(event, question.label, checkbox)}
+          checked={answer.value[question].split('|').includes(checkbox.value)}
+          onChange={(event) => onChange(event, question, checkbox)}
           value={checkbox.value}
           classNames={{ input: checkboxClasses.fixDisabled, icon: checkboxClasses.fixDisabledIcon }}
         />
@@ -62,7 +63,7 @@ function RadioGroupComponent({
   _choices: StringOption[],
   _n: number,
   idx: number,
-  question: StringOption,
+  question: string,
   response: MatrixResponse,
   answer: { value: Record<string, string> },
   onChange: (val: string, questionKey: string) => void,
@@ -77,8 +78,8 @@ function RadioGroupComponent({
         '--input-description-size': 'calc(var(--mantine-font-size-md) - calc(0.125rem * var(--mantine-scale)))',
         flex: 1,
       }}
-      onChange={(val) => onChange(val, question.label)}
-      value={answer.value[question.label]}
+      onChange={(val) => onChange(val, question)}
+      value={answer.value[question]}
     >
       <div
         style={{
@@ -119,11 +120,9 @@ export function MatrixInput({
 
   const {
     answerOptions,
-    questionOptions,
     prompt,
     secondaryText,
     required,
-    questionOrder,
   } = response;
 
   const _choiceStringToColumns: Record<string, string[]> = {
@@ -134,21 +133,10 @@ export function MatrixInput({
   };
 
   const _choices = typeof answerOptions === 'string' ? _choiceStringToColumns[answerOptions].map((entry) => ({ value: entry, label: entry })) : answerOptions.map((option) => (typeof option === 'string' ? { value: option, label: option } : option));
-  const _questions = useMemo(() => questionOptions.map((option) => (typeof option === 'string' ? { value: option, label: option } : option)), [questionOptions]);
-  const [orderedQuestions, setOrderedQuestions] = useState<StringOption[]>([]);
 
-  useEffect(() => {
-    if (questionOrder === 'random') {
-      const shuffled = [..._questions]
-        .map((value) => ({ value, sort: Math.random() }))
-        .sort((a, b) => a.sort - b.sort)
-        .map(({ value }) => value);
-      setOrderedQuestions(shuffled);
-    } else {
-      // questionOrder === 'fixed'
-      setOrderedQuestions(_questions);
-    }
-  }, [questionOrder, _questions]);
+  const { questionOrders } = useStoredAnswer();
+  const orderedQuestions = useMemo(() => questionOrders[response.id] || response.questionOptions, [questionOrders, response.id, response.questionOptions]);
+
   // Re-define on change functions. Dispatch answers to store.
   const onChangeRadio = (val: string, questionKey: string) => {
     const payload = {
@@ -254,7 +242,7 @@ export function MatrixInput({
               miw={140}
               maw={400}
             >
-              {entry.label}
+              {entry}
             </Text>
           ))}
         </div>
