+++ src/store/store.tsx
@@ -12,7 +12,7 @@ import {
 import { getSequenceFlatMap } from '../utils/getSequenceFlatMap';
 import { REVISIT_MODE } from '../storage/engines/StorageEngine';
 import { studyComponentToIndividualComponent } from '../utils/handleComponentInheritance';
-import { randomizeOptions } from '../utils/handleRandomOptions';
+import { randomizeOptions, randomizeQuestionOrder } from '../utils/handleResponseRandomization';
 
 export async function studyStoreCreator(
   studyId: string,
@@ -56,6 +56,7 @@ export async function studyStoreCreator(
           parameters: Object.hasOwn(componentConfig, 'parameters') ? (componentConfig as any).parameters : {},
           correctAnswer: Object.hasOwn(componentConfig, 'correctAnswer') ? componentConfig.correctAnswer! : [],
           optionOrders: randomizeOptions(componentConfig),
+          questionOrders: randomizeQuestionOrder(componentConfig),
         } as StoredAnswer,
       ];
     }).filter((ans) => ans !== null));
@@ -171,6 +172,7 @@ export async function studyStoreCreator(
           parameters: payload.parameters || ('parameters' in componentConfig ? componentConfig.parameters : {}) || {},
           correctAnswer: payload.correctAnswer || componentConfig.correctAnswer || [],
           optionOrders: randomizeOptions(componentConfig),
+          questionOrders: randomizeQuestionOrder(componentConfig),
         } as StoredAnswer;
         state.trialValidation[identifier] = {
           aboveStimulus: { valid: false, values: {} },
