+++ src/utils/handleResponseRandomization.ts
@@ -1,0 +1,42 @@
+import { IndividualComponent, StringOption } from '../parser/types';
+
+export function randomizeOptions(componentConfig: IndividualComponent) {
+  return componentConfig.response.reduce((acc, response) => {
+    if (
+      (response.type === 'radio' || response.type === 'checkbox' || response.type === 'buttons')
+      && response.optionOrder === 'random'
+    ) {
+      if (response.optionOrder === 'random') {
+        const options = response.options.map((option) => (typeof option === 'string' ? { value: option, label: option } : option));
+        const shuffled = [...options]
+          .map((value) => ({ value, sort: Math.random() }))
+          .sort((a, b) => a.sort - b.sort)
+          .map(({ value }) => value);
+        acc[response.id] = shuffled;
+      } else {
+        acc[response.id] = response.options.map((option) => (typeof option === 'string' ? { value: option, label: option } : option));
+      }
+    }
+    return acc;
+  }, {} as Record<string, StringOption[]>);
+}
+
+export function randomizeQuestionOrder(componentConfig: IndividualComponent) {
+  return componentConfig.response.reduce((acc, response) => {
+    if (
+      (response.type === 'matrix-radio' || response.type === 'matrix-checkbox')
+    ) {
+      if (response.questionOrder === 'random') {
+        const questions = response.questionOptions;
+        const shuffled = [...questions]
+          .map((value) => ({ value, sort: Math.random() }))
+          .sort((a, b) => a.sort - b.sort)
+          .map(({ value }) => value);
+        acc[response.id] = shuffled;
+      } else {
+        acc[response.id] = response.questionOptions;
+      }
+    }
+    return acc;
+  }, {} as Record<string, string[]>);
+}
